e2e tests: move MNP code to a separate file
